requirement.txt
aniso8601==9.0.1
autopep8==1.6.0
bidict==0.22.0
boto3==1.22.1
botocore==1.25.1
certifi==2021.10.8
Flask==2.0.2
Flask-MySQLdb==0.2.0
flask-paginate==2022.1.8
Jinja2==3.0.3
jmespath==1.0.0
mysql-connector-python==8.0.28
mysqlclient==2.1.0
pycodestyle==2.8.0
pycparser==2.21
pytz==2021.3
PyYAML==6.0
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.2
wsproto==1.1.0
